% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{C_Elegans_data}
\alias{C_Elegans_data}
\title{C_elegans_data set}
\format{
a data frame with 8 rows ans 6 columns
}
\source{
an example data frame
}
\usage{
C_Elegans_data
}
\description{
C_elegans_data set
}
\keyword{datasets}
