% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Change_type.R
\name{Change_type}
\alias{Change_type}
\title{change column data type}
\usage{
Change_type(file, file_col_1, type_1, file_col_2, type_2, file_col_3, type_3)
}
\arguments{
\item{file}{the file / path to file, that needs to be changed}

\item{file_col_1}{the name of the column to change}

\item{type_1}{the type of data it should be changed to ("dbl" , "factor" , "character")}

\item{file_col_2}{the name of the column to change}

\item{type_2}{the type of data it should be changed to ("dbl" , "factor" , "character")}

\item{file_col_3}{the name of the column to change}

\item{type_3}{the type of data it should be changed to ("dbl" , "factor" , "character")}
}
\value{
the new data file with the correct data types
}
\description{
change column data type
}
\examples{
data_2 <-Change_type("inst", "comp", "dbl", "RawData", "dbl", "compName", "factor")
}
