% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_veh_con.R
\name{Plot_veh_con}
\alias{Plot_veh_con}
\title{Plot the control samples}
\usage{
Plot_veh_con(file, col_neg, name_pos, name_veh, name_neg, RawData)
}
\arguments{
\item{file}{file with sample data}

\item{col_neg}{column with names control samples}

\item{name_pos}{name of positive control}

\item{name_veh}{name of vehicle control}

\item{name_neg}{name of negative control}

\item{RawData}{column with data}
}
\value{
plot with with all the averages per control
}
\description{
Plot the control samples
}
\examples{
\dontrun{Plot_veh_con(data_t2, expType,"control" ,"control", "control",RawData)}
}
