% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_and_store.R
\name{Read_and_store}
\alias{Read_and_store}
\title{download data}
\usage{
Read_and_store(pathname_of_file, file_type)
}
\arguments{
\item{pathname_of_file}{a vector or string that contains a path to the data file}

\item{file_type}{the type of data file used (.xlsx , .csv or .csv2)}
}
\value{
the newly downloaded file
}
\description{
download data
}
\examples{
\dontrun{data  <- Read_and_store("inst/CE.LIQ.FLOW.062_Tidydata.xlsx" , ".xlsx")}
}
